Summarizing all the styles of the page
