+++ src/common/Header/index.js
@@ -1,9 +1,7 @@
 import { Title } from "./styled";
 
 const Header = ({ title }) => (
-    <header>
-        <Title>{title}</Title>
-    </header>
+    <Title>{title}</Title>
 );
 
 export default Header;
