+++ src/common/Header/styled.js
@@ -1,8 +1,19 @@
 import styled from "styled-components";
 
-export const Title = styled.h1`
-  margin: 0;
+export const Title = styled.span`
   font-size: 36px;
   font-weight: 600;
-  word-break: break-word;
+  line-height: 43px;
+  grid-column: 1 / -1;
+  color: ${({ theme }) => theme.colors.woodsmoke};
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletS}) {
+    font-size: 28px;
+    line-height: 34px;
+  }
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.mobileM}) {
+    font-size: 20px;
+    line-height: 24px;
+  }
 `;
