+++ src/common/MoviesTilesList/styled.js
@@ -1,43 +1,134 @@
 import styled from "styled-components";
+import { ReactComponent as StarIcon } from '../../image/Vector.svg';
 
-export const StyledTileList = styled.section`
-    display: grid;
-    grid-gap: 32px;
+export const MoviePoster = styled.img`
+  border-radius: 5px;
 
-    @media (max-width: 450px) {
-        grid-gap: 12px;
-    };
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    width: 114px;
+    height: 169px;
+  }
 `;
 
-export const ListHeader = styled.header`
-    font-size: 36px;
-    font-weight: 600;
+export const StyledTile = styled.div`
+  background-color: ${({ theme }) => theme.colors.white};
+  box-shadow: rgba(186, 199, 213, 0.5) 0px 4px 12px;
+  display: flex;
+  flex-direction: column;
+  padding: 16px;
+  width: 100%;
+  position: relative;
+  gap: 16px;
+  cursor: pointer;
 
-    @media (max-width: 450px) {
-        font-size: 20px;
-    };
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    flex-direction: row;
+  }
 `;
 
-export const ListSection = styled.ul`
-    display: grid;
-    grid-template-columns: repeat(4, 25%);
-    align-items: start;
+export const MovieInfo = styled.div`
+  display: flex;
+  flex-direction: column;
+  flex: 1 1 0%;
+`;
 
-    grid-gap: 24px;
-    padding-left: 0px;
+export const MovieTitle = styled.h2`
+  font-size: 22px;
+  font-weight: 500;
+  line-height: 29px;
+  margin: 0 0 8px;
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    font-size: 16px;
+    line-height: 21px;
     margin: 0;
+  }
+`;
 
-    @media (max-width: 900px) {
-        grid-template-columns: repeat(3, 31%);
-    };
+export const MovieSubTitle = styled.div`
+  font-size: 16px;
+  line-height: 25px;
+  color: ${({ theme }) => theme.colors.waterloo};
+  margin-bottom: 8px;
 
-    @media (max-width: 700px) {
-        grid-template-columns: repeat(2, 50%);
-        justify-content: center;
-        grid-gap: 16px;
-    };
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    font-size: 13px;
+    margin-bottom: 0;
+  }
+`;
 
-    @media (max-width: 450px) {
-        grid-template-columns: 100%;
-    };
+export const MovieTags = styled.ul`
+  display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
+  list-style: none;
+  margin: 0 0 8px;
+  padding: 0;
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    margin: 6px 0 8px;
+  }
+`;
+
+export const MovieTag = styled.li`
+  padding: 8px 16px;
+  background: ${({ theme }) => theme.colors.mystic};
+  border-radius: 5px;
+  font-size: 14px;
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    padding: 4px 8px;
+    font-size: 10px;
+  }
+`;
+
+export const RatingWrapper = styled.div`
+  display: flex;
+  gap: 12px;
+  margin-top: auto;
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    margin-top: 8px;
+  }
+`;
+
+export const Rating = styled.div`
+  display: flex;
+  gap: 12px;
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    gap: 8px;
+  }
+`;
+
+export const VoteCount = styled.div`
+  font-size: 16px;
+  line-height: 24px;
+  color: ${({ theme }) => theme.colors.waterloo};
+  
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    font-size: 13px;
+    line-height: 17px;
+  }
+`;
+
+export const RatingNumber = styled.div`
+  font-size: 16px;
+  font-weight: 600;
+  line-height: 24px;
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    font-size: 13px;
+    line-height: 17px;
+  }
+`;
+
+export const RatingStar = styled(StarIcon)`
+  width: 24px;
+  height: 24px;
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    width: 16px;
+    height: 16px;
+  }
 `;
