+++ src/features/ActorList/index.js
@@ -1,10 +1,11 @@
 import { useNavigate } from "react-router-dom";
 import { actors } from './actorsData';
-import { Title, Section } from './styled';
+import { Section } from './styled';
 import { toPerson } from "../../routes";
 import { Pagination } from "../../common/Pagination";
 import { Container } from "../../common/Container";
 import { PeopleTilesList } from "../../common/PeopleTilesList"
+import Header from "../../common/Header";
 
 const ActorsList = () => {
     const navigate = useNavigate();
@@ -16,14 +17,13 @@ const ActorsList = () => {
     return (
         <Container>
             <Section>
-                <Title>Popular people</Title>
+            <Header title="Popular people" />
                 {actors.map((actor, index) => (
                     <PeopleTilesList
                         key={index}
                         onClick={() => handleActorClick(actor.id)}
                         photo={actor.photo}
                         name={actor.name}
-                        subName={actor.subname}
                     />
                 ))}
