+++ src/features/ActorPage/Content/Crew/filmsDataCrew.js
@@ -1,7 +1,7 @@
 export const films = [
     {
         name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        photo: require('../../../../image/poster2.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
@@ -9,7 +9,7 @@ export const films = [
     },
     {
         name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        photo: require('../../../../image/poster2.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
@@ -17,7 +17,7 @@ export const films = [
     },
     {
         name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        photo: require('../../../../image/poster2.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
@@ -25,7 +25,7 @@ export const films = [
     },
     {
         name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        photo: require('../../../../image/poster2.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
@@ -33,7 +33,7 @@ export const films = [
     },
     {
         name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        photo: require('../../../../image/poster2.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
@@ -41,7 +41,7 @@ export const films = [
     },
     {
         name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        photo: require('../../../../image/poster2.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
@@ -49,7 +49,7 @@ export const films = [
     },
     {
         name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        photo: require('../../../../image/poster2.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
@@ -57,7 +57,7 @@ export const films = [
     },
     {
         name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        photo: require('../../../../image/poster2.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
