+++ src/features/MovieDetailsPage/components/Crew/styled.js
@@ -11,6 +11,7 @@ export const Title = styled.h1`
   @media (max-width: ${({ theme }) => theme.breakpoints.tabletS}) {
     line-height: 34px;
     font-size: 28px;
+    font-size: 32px;
   }
   @media (max-width: ${({ theme }) => theme.breakpoints.mobileM}) {
     line-height: 24px;
@@ -23,12 +24,12 @@ export const Section = styled.section`
   display: grid;
   grid-template-columns: repeat(6, minmax(100px, 1fr));
   gap: 24px;
+  width: 100%;
   justify-content: center;
 
   @media (max-width: ${({ theme }) => theme.breakpoints.tabletL}) {
     grid-template-columns: repeat(5, minmax(100px, 1fr));
     gap: 20px;
-    width: 100%;
   }
 
   @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
