+++ src/features/MovieDetailsPage/components/Details/index.js
@@ -1,0 +1,14 @@
+import { DetailsTile } from "../../../../common/DetailsTile";
+import { GenresList } from "../../../../common/GenresList";
+import { Rates } from "../../../../common/Rates";
+
+export const Details = () => (
+    <DetailsTile
+        extraContent={
+            <>
+                <GenresList />
+                <Rates />
+            </>
+        }
+    />
+);
