--- src/features/MoviesListPage/filmsData.js
@@ -1,42 +1,0 @@
-export const films = [
-    {
-        name: "Mulan",
-        photo: "path_to_photo_2",
-        year: "2020",
-        genres: ["Action", "Adventure", "Drama"],
-        rate: "7,8",
-        vote: "35 votes"
-    },
-    {
-        name: "Mulan",
-        photo: "path_to_photo_2",
-        year: "2020",
-        genres: [],
-        rate: "7,8",
-        vote: "35 votes"
-    },
-    {
-        name: "Mulan",
-        photo: "path_to_photo_2",
-        year: "2020",
-        genres: ["Action"],
-        rate: "7,8",
-        vote: "35 votes"
-    },
-    {
-        name: "Mulan long title long title Mulan long title long",
-        photo: "path_to_photo_2",
-        year: "2020",
-        genres: ["Action"],
-        rate: "7,8",
-        vote: "35 votes"
-    },
-    {
-        name: "Mulan",
-        photo: "path_to_photo_2",
-        year: "",
-        genres: ["Action"],
-        rate: "7,8",
-        vote: "35 votes"
-    },
-];
