+++ src/features/MoviesListPage/moviesData.js
@@ -1,67 +1,66 @@
 export const films = [
     {
-        name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        name: "Mulan",
+        photo: require('../../image/poster.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
         vote: "35 votes"
     },
     {
-        name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        name: "Mulan",
+        photo: require('../../image/poster.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
         vote: "35 votes"
     },
     {
-        name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        name: "Mulan",
+        photo: require('../../image/poster.png'),
+        year: "Mulan (2020)",
+        genres: ["Action"],
+        rate: "7,8",
+        vote: "35 votes"
+    },
+    {
+        name: "Mulan",
+        photo: require('../../image/poster.png'),
+        year: "Mulan (2020)",
+        genres: ["Action"],
+        rate: "7,8",
+        vote: "35 votes"
+    },
+    {
+        name: "Mulan",
+        photo: require('../../image/poster.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
         vote: "35 votes"
     },
     {
-        name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        name: "Mulan",
+        photo: require('../../image/poster.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
         vote: "35 votes"
     },
     {
-        name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        name: "Mulan",
+        photo: require('../../image/poster.png'),
         year: "Mulan (2020)",
-        genres: ["Action", "Adventure", "Drama"],
+        genres: ["Action"],
         rate: "7,8",
         vote: "35 votes"
     },
     {
-        name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
+        name: "Mulan",
+        photo: require('../../image/poster.png'),
         year: "Mulan (2020)",
         genres: ["Action", "Adventure", "Drama"],
         rate: "7,8",
         vote: "35 votes"
     },
-    {
-        name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
-        year: "Mulan (2020)",
-        genres: ["Action", "Adventure", "Drama"],
-        rate: "7,8",
-        vote: "35 votes"
-    },
-    {
-        name: "Chloe Coleman",
-        photo: require('../../../../image/poster.png'),
-        year: "Mulan (2020)",
-        genres: ["Action", "Adventure", "Drama"],
-        rate: "7,8",
-        vote: "35 votes"
-    },
-
 ];
