+++ src/features/MoviesListPage/styled.js
@@ -1,10 +1,29 @@
 import styled from "styled-components";
 
-export const TileContainer = styled.div`
-  margin-top: 24px;
+export const TileContainer = styled.section`
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-  column-gap: 24px;
-  row-gap: 24px;
-  justify-content: space-between;
+  grid-template-columns: repeat(4, 324px);
+  grid-gap: 24px;
+  margin: 64px auto;
+  width: 1400px;
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.laptop}) {
+    grid-template-columns: repeat(4, minmax(200px, 1fr));
+    width: 100%;
+  }
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletL}) {
+    grid-template-columns: repeat(3, minmax(200px, 1fr));
+    width: 100%;
+  }
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletM}) {
+    grid-template-columns: repeat(2, 1fr);
+  }
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.tabletS}) {
+    grid-template-columns: repeat(1,  1fr);
+    margin-top: 24px;
+    grid-gap: 16px;
+  }
 `;
